included exact propagation. How?
each signal contains two columns. the first are the mean values, the second are the variances at each time instant k.
if we want to predict at the specified time instant k, the "predictAt" method check if variance. If the variances of regressor vectors are non-zero, the exact propagation is done. Otherwise, if vairance are exactly zero, naive prediction is done.

SignalsModel was also suitably modified to incorporate signals with variance. waiting for bugs...